auth/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.37.2
pydantic>=2.0.0,<3.0.0
pydantic-settings==2.1.0
motor==3.3.2
pymongo==4.6.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0
